Minor fixes

+++ comm/source/mnWinCommStreams.pas
@@ -131,8 +131,7 @@ begin
         DCB.Flags := DCB.Flags or dcb_OutxCTSFlow;
       if OutDSRFlow then
         DCB.Flags := DCB.Flags or dcb_OutxDSRFlow;
-      DCB.Flags := DCB.Flags or CControlDTR[ControlDTR]
-        or CControlRTS[ControlRTS];
+      DWORD(DCB.Flags) := DWORD(DCB.Flags) or CControlDTR[ControlDTR] or CControlRTS[ControlRTS];
       if XonXoffOut then
         DCB.Flags := DCB.Flags or dcb_OutX;
       if XonXoffIn then
